Корректировка отображения диаграммы

+++ src/dia/node.h
@@ -69,8 +69,8 @@ public:
     bool advance();
 
     NodeInfo *nodeInfo;
-    int x;
-    int y;
+    qreal x;
+    qreal y;
 
     QRectF boundingRect() const Q_DECL_OVERRIDE;
    // QPainterPath shape() const Q_DECL_OVERRIDE;
